Add files via upload
Google PlayStore App Analysis Using Python. I imported my data 
from kaggle and using the famous pandas library to read in my 
csv dataset into jupyter notebook. In performing the analysis 
I intended to answer the followinhg the tasks.
While working on question 11, I encountered a major issue because
I cant get the average of 'object' datatype, I tried to convert 
it to either int or float dtype but still encountered a problem
('Life Made WI-Fi Touchscreen Photo Frame' cointains 3.0M as a 
review). I have to use the "replace" function to remove the "M"
by replacing it with "3.0" instead of "3.0M" almost same was 
applicable to question 16 but in this case it was the case of 
having ',' and '+' in the dataset. I later found out that there 
is a row having 'Free'

1. Show the TOP 20 rows of the dataset
2. Show the BOTTOM 5 rows of the dataset
3. Show the number of rows and number of columns in the dataset
4. Getting more information about the dataset to enable 
accurate analysis.
5. Show overall statistics about the the dataset for both 
numerical and non-numerical values.
6. Total number of apps titles that contain Python, R and 
Programinng.
7. Find the average app rating
8. Find the total number of unique catagory
9. Find the category with the highest average rating
10. Find the total number of apps with 5 star rating
11. Find the average values of reviews
12. Find the total number of Free and Paid Apps
13. Which App has the maximum reviews
14. Display Top 10 apps having highest reviews
15. Find the average rating of free and paid apps
16. Display Top 10 app having maximum installs